split facade single endpoint generator into own file to keep tidy
